% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{gwas_download_files}
\alias{gwas_download_files}
\title{Convenient wrapper for downloading summary statistics files}
\usage{
gwas_download_files(file_urls, output_paths, max_concurrent = 4)
}
\arguments{
\item{file_urls}{Vector of file URLs to download}

\item{output_paths}{Vector of output paths (must match length of file_urls)}

\item{max_concurrent}{Maximum number of concurrent downloads (default: 4)}
}
\value{
Status message with download results
}
\description{
Convenient wrapper for downloading summary statistics files
}
