% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{gwas_filter}
\alias{gwas_filter}
\title{Create a filter object for GWAS queries}
\usage{
gwas_filter(
  p_value = NULL,
  bp_location = NULL,
  study = NULL,
  trait = NULL,
  reveal = NULL,
  start = NULL,
  size = NULL
)
}
\arguments{
\item{p_value}{Optional p-value range as c(min, max)}

\item{bp_location}{Optional base pair location range as c(min, max)}

\item{study}{Optional study accession}

\item{trait}{Optional trait ID}

\item{reveal}{Optional reveal mode ("raw" or "all")}

\item{start}{Optional offset number (default: 0)}

\item{size}{Optional number of items returned (default: 20)}
}
\value{
A list representing filter parameters
}
\description{
Create a filter object for GWAS queries
}
